requirements.txt
pandas-ta-openbb
agno
openai 
alpaca-py 
requests
yfinance
python-dotenv